add: admin panel decorators
